+++ allocator/srcs/ft_realloc.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_realloc.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: gcrisp <marvin@42.fr>                      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/02/24 12:25:09 by gcrisp            #+#    #+#             */
+/*   Updated: 2025/02/24 12:25:10 by gcrisp           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "allocator.h"
 
 static void	*contract_alloc(t_dllist *chunks, size_t new_size)
@@ -7,7 +19,8 @@ static void	*contract_alloc(t_dllist *chunks, size_t new_size)
 	difference = ((t_chunk *)chunks->data)->size - new_size;
 	((t_chunk *)chunks->data)->size = new_size;
 	((t_chunk *)chunks->next->data)->size += difference;
-	((t_chunk *)chunks->next->data)->start = (size_t)((t_chunk *)chunks->next->data)->start - difference;
+	((t_chunk *)chunks->next->data)->start
+		= ((t_chunk *)chunks->next->data)->start - difference;
 	return (((t_chunk *)chunks->data)->start);
 }
 
@@ -18,7 +31,8 @@ static void	*expand_alloc(t_dllist *chunks, size_t new_size)
 	difference = new_size - ((t_chunk *)chunks->data)->size;
 	((t_chunk *)chunks->data)->size = new_size;
 	((t_chunk *)chunks->next->data)->size -= difference;
-	((t_chunk *)chunks->next->data)->start = (size_t)((t_chunk *)chunks->next->data)->start + difference;
+	((t_chunk *)chunks->next->data)->start
+		= ((t_chunk *)chunks->next->data)->start + difference;
 	return (((t_chunk *)chunks->data)->start);
 }
 
@@ -27,7 +41,8 @@ static void	*move_alloc(t_dllist *chunks, size_t new_size)
 	void	*out;
 
 	out = ft_malloc(new_size);
-	ft_memcpy(out, ((t_chunk *)chunks->data)->start, ((t_chunk *) chunks->data)->size);
+	ft_memcpy(out, ((t_chunk *)chunks->data)->start,
+		((t_chunk *) chunks->data)->size);
 	free_chunk(&chunks);
 	return (out);
 }
@@ -50,7 +65,8 @@ void	*ft_realloc(void *ptr, size_t new_size)
 		return (ptr);
 	if (new_size < ((t_chunk *)chunks->data)->size)
 		return (contract_alloc(ptr, new_size));
-	if (((t_chunk *)chunks->data)->size + ((t_chunk *)chunks->next->data)->size <= new_size)
+	if (((t_chunk *)chunks->data)->size + ((t_chunk *)chunks->next->data)->size
+		<= new_size)
 		return (expand_alloc(chunks, new_size));
 	return (move_alloc(chunks, new_size));
 }
